requirements.txt
click==8.3.1
cloudpickle==3.1.2
ConfigSpace==1.2.1
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
cycler==0.12.1
dask==2026.1.2
dask-jobqueue==0.9.0
distributed==2026.1.2
emcee==3.1.6
filelock==3.20.3
fonttools==4.61.1
fsspec==2026.2.0
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
lap==0.5.12
locket==1.0.0
MarkupSafe==3.0.3
matplotlib==3.9.2
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
networkx==3.6.1
numpy==2.1.1
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
opencv-python==4.10.0.84
packaging==26.0
pandas==2.2.2
partd==1.4.2
patsy==1.0.2
pillow==10.4.0
pluggy==1.6.0
psutil==7.2.2
pycocotools==2.0.11
pynisher==1.0.10
pyparsing==3.3.2
pyrfr==0.9.0
pytest==8.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2026.1.15
scikit-learn==1.5.2
scipy==1.17.0
setuptools==80.10.2
six==1.17.0
smac==2.3.0
sortedcontainers==2.4.0
statsmodels==0.14.3
sympy==1.14.0
tabulate==0.9.0
tblib==3.2.2
threadpoolctl==3.6.0
toolz==1.1.0
torch==2.10.0
torchvision==0.25.0
tornado==6.5.4
tqdm==4.66.5
triton==3.6.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
zict==3.0.0
